Not finished

+++ src/app/register/register.component.ts
@@ -1,7 +1,6 @@
 import { Component } from '@angular/core';
 import { EmailValidator, FormBuilder, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import { catchError } from 'rxjs';
 
 @Component({
   selector: 'app-register',
@@ -10,11 +9,12 @@ import { catchError } from 'rxjs';
 })
 export class RegisterComponent {
   registerForm = this.fb.group({
-    name:[''],
-    surname:[''],
-    phone:[''],
+    name:['', [Validators.required,]],
+    surname:['', [Validators.required,]],
+    phone:['', [Validators.required,]],
     email:['', [Validators.required, Validators.email]],
-    password:['', [Validators.minLength(5)]]
+    // docNumber:['', Validators.required],
+    password:['', [Validators.required, Validators.minLength(5)]]
   })
   constructor(
     public route: Router,
@@ -27,28 +27,34 @@ export class RegisterComponent {
   saveUser(){    
     let user = this.registerForm.value;
     let db = JSON.parse(JSON.stringify(localStorage.getItem('db')));
-    console.log(db);
-    if(db===null){ // If there's no db, creates a new one and saves it into local storage
-      console.log("is null");
-      db=[];
-      let toPushUser = JSON.stringify(user);
-      db[db.length]=toPushUser;
-      localStorage.clear();
-      localStorage.setItem('db', db);
+    // Verifies that there's no empty field
+    if((user.name==null)||(user.surname==null)||(user.phone==null)||(user.email==null)||(user.password==null)){
+      if(db===null){ // If there's no db, creates a new one and saves it into local storage
+        console.log("is null");
+        db=[];
+        user.email?.toLowerCase();
+        console.log(user.email);
+        let toPushUser = JSON.stringify(user);
+        db[db.length]=toPushUser;
+        localStorage.clear();
+        localStorage.setItem('db', db);
+      }else{
+        let dbLocalStorage = localStorage.getItem('db');
+        let toCompare = JSON.parse(JSON.stringify(dbLocalStorage));
+         // Verifies if there's any email associated in the 'db'
+         user.email?.toLowerCase();
+          if (toCompare.includes(user.email)){  
+            console.error("Email duplicado");
+          }else{ // If not registered saves it into the 'db'
+          db = [dbLocalStorage];
+          db[db.length]=JSON.stringify(user);
+          localStorage.clear();
+          localStorage.setItem('db',db);
+        }}
     }else{
-      let dbLocalStorage = localStorage.getItem('db');
-      let toCompare = JSON.parse(JSON.stringify(dbLocalStorage));
-       // Verifies if there's any email associated in the 'db'
-      if (toCompare.includes(user.email)){
-        console.error("Email duplicado");
-      }else{ // If not registered saves it into the 'db'
-      db = [dbLocalStorage];
-      db[db.length]=JSON.stringify(user);
-      localStorage.clear();
-      localStorage.setItem('db',db);
-    }
-    }
-  }
+      alert("llene todos los campos");
+      window.location.reload();
+}}
   get name(){
     return this.registerForm.controls['name'];
   }
